requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.0.3
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
